Fix getKey handling of falsy pk and sk
getKey checks for the presence of pk and sk and
throws an error if they are not present when they should
be present. The code errors when the keys are falsy.
This is a problem for falsy values like 0 for numeric
types which are allowed in dynamo.

Updated the code to allow 0 but disallow undefined, null
and ''.