Validate that removed fields are not required
Fixes #255
Defining a table attribute with `required: true` should mean that values always exist for that field (barring extraneous factors like schema changes and writes that may not go through this library). Updates may or may not specify a new value, but that new value should not be null, and the field cannot be removed outright.

This fixes two cases of this bug:
1) You can remove a `required: true` value by passing $remove: attr or `$remove: [attr]`
2) You can remove a `required: true` value by passing `attr: ''` if `removeNullAttributes` is set.